Example model output data representing mean point forecasts

These are based on a subset of forecasts of daily incident hospital admissions for COVID-19 that were submitted to the US COVID-19 Forecast hub, which have been lightly modified for expository purposes. To obtain point forecasts, we used the distfromq package for R to estimate the quantile function, drew 100,000 samples from the distribution using inverse transform sampling, and computed the mean of those samples.

Usage

example_mean_model_output

Format

example_mean_model_output

A data frame with 16 rows and 8 columns:

origin_date

Date: Reference date for the forecast.

horizon

integer: Forecast horizon, in days.

location

character: FIPS codes identifying the location being forecasted.

target

character: The forecast target. All values are "inc covid hosp".

output_type

character: Output type used to represent the forecast. All values are "mean".

output_type_id

NA: This column is not used when the output type is "mean". All values are "NA".

value

numeric: The mean of the forecast distribution.

model_id

character: Identifier for the model that produced the forecast.

Source

https://github.com/reichlab/covid19-forecast-hub