Example target data recording outcomes that are predicted in the forecasts stored in example_quantile_model_output.

These are daily incident hospital admissions for COVID-19 that were reported on HealthData.gov. They are organized so as to match the forecast task id columns in example_quantile_model_output, so that the observed data can be joined into the forecast data frame correctly.

Usage

example_target_data

Format

example_target_data

A data frame with 8 rows and 5 columns:

origin_date

Date: Reference date for the forecast.

horizon

integer: Forecast horizon, in days.

location

character: FIPS codes identifying the location being forecasted.

target

character: The forecast target. All values are "inc covid hosp".

value

integer: The number of reported hospitalizations.

Source

https://healthdata.gov/Hospital/COVID-19-Reported-Patient-Impact-and-Hospital-Capa/g62h-syeh/about_data